+++ disparity_view/__init__.py
@@ -4,4 +4,4 @@ from .depth_to_normal import DepthToNormalMap
 from .o3d_project import project_from_left_and_disparity, gen_right_image
 from .o3d_project import make_animation_gif
 
-from .zed_info import CameraParameter
+from .cam_param import CameraParameter
