+++ scripts/read_camera_json.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from disparity_view.zed_info import CameraParameter
+from disparity_view.cam_param import CameraParameter
 
 json_file = Path("../test/zed-imgs/camera_param.json")
 camera_parameter = CameraParameter.load_json(json_file)
